+++ docs/changelog.md
@@ -27,6 +27,8 @@
 - Fixed an issue that caused the developer announcements to be shown to often.
 - Fixed GCode console history resetting after leaving the page.
 - Fixed full-screen webcam page not opening correctly from the overview page.
+- Fixed flickering of overview page printer cards while printing.
+-
 
 ## [2.7.0] - 2024-06-26